HW Refactoring: Extract classes to separate files
