+++ hotwallet/lib/HotWalletRedis.js
@@ -1,0 +1,78 @@
+const { promisify } = require("util")
+const HotWallet = require("./HotWallet").HotWallet
+
+class HotWalletRedis {
+  constructor(envs, redisClient) {
+    this.envs = envs
+    this.client = redisClient
+
+    // Set error handler
+    this.client.on("error", function (err) {
+      console.error(`Redis client error: ${err}`);
+    })
+  }
+
+  walletKeyName() {
+    return `wallet_for_project_${this.envs.projectId}`
+  }
+
+  async hotWallet() {
+    const savedHW = await this.hgetall(this.walletKeyName())
+
+    if (savedHW) {
+      const optedInApps = JSON.parse(savedHW.optedInApps || "[]")
+
+      return new HotWallet(savedHW.address, savedHW.mnemonic, optedInApps)
+    } else {
+      return undefined
+    }
+  }
+
+  async hotWalletAddress() {
+    return await this.hget(this.walletKeyName(), "address")
+  }
+
+  async hotWalletMnenonic() {
+    return await this.hget(this.walletKeyName(), "mnemonic")
+  }
+
+  async isHotWalletCreated() {
+    return (await this.hotWallet()) !== undefined
+  }
+
+  async saveNewHotWallet(wallet) {
+    await this.hset(this.walletKeyName(), "address", wallet.address, "mnemonic", wallet.mnemonic)
+    console.log(`Keys for a new hot wallet has been saved into ${this.walletKeyName()}`)
+    return true
+  }
+
+  async saveOptedInApps(optedInApps) {
+    optedInApps = JSON.stringify(optedInApps)
+    await this.hset(this.walletKeyName(), "optedInApps", optedInApps)
+    console.log(`Opted-in apps (${optedInApps}) has been saved into ${this.walletKeyName()}`)
+    return true
+  }
+
+  async deleteCurrentKey() {
+    await this.del(this.walletKeyName())
+    console.log(`Wallet keys has been deleted: ${this.walletKeyName()}`)
+    return true
+  }
+
+  async hset(...args) {
+    return await (promisify(this.client.hset).bind(this.client))(...args)
+  }
+
+  async hget(...args) {
+    return await (promisify(this.client.hget).bind(this.client))(...args)
+  }
+
+  async hgetall(...args) {
+    return await (promisify(this.client.hgetall).bind(this.client))(...args)
+  }
+
+  async del(...args) {
+    return await (promisify(this.client.del).bind(this.client))(...args)
+  }
+}
+exports.HotWalletRedis = HotWalletRedis
