+++ hotwallet/lib/hotwalletUtils.js
@@ -1,390 +1,14 @@
-const { promisify } = require("util")
 const algosdk = require("algosdk")
 const chainjs = require("@open-rights-exchange/chainjs")
-const axios = require("axios")
-const BigNumber = require('bignumber.js')
 
-class HotWallet {
-  constructor(address, mnemonic, optedInApps = []) {
-    this.address = address
-    this.mnemonic = mnemonic
-    this.optedInApps = optedInApps
-  }
-
-  isOptedInToApp(appIndexToCheck) {
-    if (!this.optedInApps) { return false }
-    return this.optedInApps.includes(appIndexToCheck)
-  }
-
-  secretKey() {
-    const { sk } = algosdk.mnemonicToSecretKey(this.mnemonic)
-    return sk
-  }
-}
+// Custom classes
+const HotWallet = require("./HotWallet").HotWallet
 exports.HotWallet = HotWallet
-
-class HotWalletRedis {
-  constructor(envs, redisClient) {
-    this.envs = envs
-    this.client = redisClient
-
-    // Set error handler
-    this.client.on("error", function (err) {
-      console.error(`Redis client error: ${err}`);
-    })
-  }
-
-  walletKeyName() {
-    return `wallet_for_project_${this.envs.projectId}`
-  }
-
-  async hotWallet() {
-    const savedHW = await this.hgetall(this.walletKeyName())
-
-    if (savedHW) {
-      const optedInApps = JSON.parse(savedHW.optedInApps || "[]")
-
-      return new HotWallet(savedHW.address, savedHW.mnemonic, optedInApps)
-    } else {
-      return undefined
-    }
-  }
-
-  async hotWalletAddress() {
-    return await this.hget(this.walletKeyName(), "address")
-  }
-
-  async hotWalletMnenonic() {
-    return await this.hget(this.walletKeyName(), "mnemonic")
-  }
-
-  async isHotWalletCreated() {
-    return (await this.hotWallet()) !== undefined
-  }
-
-  async saveNewHotWallet(wallet) {
-    await this.hset(this.walletKeyName(), "address", wallet.address, "mnemonic", wallet.mnemonic)
-    console.log(`Keys for a new hot wallet has been saved into ${this.walletKeyName()}`)
-    return true
-  }
-
-  async saveOptedInApps(optedInApps) {
-    optedInApps = JSON.stringify(optedInApps)
-    await this.hset(this.walletKeyName(), "optedInApps", optedInApps)
-    console.log(`Opted-in apps (${optedInApps}) has been saved into ${this.walletKeyName()}`)
-    return true
-  }
-
-  async deleteCurrentKey() {
-    await this.del(this.walletKeyName())
-    console.log(`Wallet keys has been deleted: ${this.walletKeyName()}`)
-    return true
-  }
-
-  async hset(...args) {
-    return await (promisify(this.client.hset).bind(this.client))(...args)
-  }
-
-  async hget(...args) {
-    return await (promisify(this.client.hget).bind(this.client))(...args)
-  }
-
-  async hgetall(...args) {
-    return await (promisify(this.client.hgetall).bind(this.client))(...args)
-  }
-
-  async del(...args) {
-    return await (promisify(this.client.del).bind(this.client))(...args)
-  }
-}
+const HotWalletRedis = require("./HotWalletRedis").HotWalletRedis
 exports.HotWalletRedis = HotWalletRedis
-
-class AlgorandBlockchain {
-  constructor(envs) {
-    this.envs = envs
-    this.blockchainNetwork = envs.blockchainNetwork
-    const endpoints = this.endpointsByNetwork(this.blockchainNetwork)
-    this.algoChain = new chainjs.ChainFactory().create(chainjs.ChainType.AlgorandV1, endpoints)
-    // It is necessary to cache values gotten from blockchain
-    this.optedInApps = {}
-    this.algoBalances = {}
-    this.tokenBalances = {}
-  }
-
-  algoMainnetEndpoints() {
-    return [{
-      url: 'https://mainnet-algorand.api.purestake.io/ps2',
-      options: { indexerUrl: 'https://mainnet-algorand.api.purestake.io/idx2', headers: [{ 'x-api-key': this.envs.purestakeApi }] },
-    }]
-  }
-
-  algoTestnetEndpoints() {
-    return [{
-      url: 'https://testnet-algorand.api.purestake.io/ps2',
-      options: { indexerUrl: 'https://testnet-algorand.api.purestake.io/idx2', headers: [{ 'x-api-key': this.envs.purestakeApi }] },
-    }]
-  }
-
-  algoBetanetEndpoints() {
-    return [{
-      url: 'https://betanet-algorand.api.purestake.io/ps2',
-      options: { indexerUrl: 'https://betanet-algorand.api.purestake.io/idx2', headers: [{ 'x-api-key': this.envs.purestakeApi }] },
-    }]
-  }
-
-  endpointsByNetwork(network) {
-    switch (network) {
-      case 'algorand_test':
-        return this.algoTestnetEndpoints()
-      case 'algorand_beta':
-        return this.algoBetanetEndpoints()
-      case 'algorand':
-        return this.algoMainnetEndpoints()
-      default:
-        console.error("Unknown or unsupported network")
-    }
-  }
-
-  generateAlgorandKeyPair() {
-    const account = algosdk.generateAccount()
-    const mnemonic = algosdk.secretKeyToMnemonic(account.sk);
-
-    return new HotWallet(account.addr, mnemonic)
-  }
-
-  async connect() {
-    if (!this.algoChain.isConnected) {
-      await this.algoChain.connect()
-    }
-  }
-
-  async getAppLocalState(hotWalletAddress, appIndex) {
-    await this.connect()
-
-    try {
-      const hwAccountDetails = await this.algoChain.algoClientIndexer.lookupAccountByID(hotWalletAddress).do()
-      return hwAccountDetails.account['apps-local-state'].filter(app => !app.deleted && app.id === appIndex)[0]
-    } catch (err) {
-      return {}
-    }
-  }
-
-  async getTokenBalance(hotWalletAddress) {
-    if (hotWalletAddress in this.tokenBalances) { return this.tokenBalances[hotWalletAddress] }
-
-    const localState = (await this.getAppLocalState(hotWalletAddress, this.envs.optInApp))
-    const values = localState["key-value"]
-    if (!values) { return 0 }
-
-    const balanceValueArray = values.filter(val => val.key === "YmFsYW5jZQ==") // balance
-    if (balanceValueArray.length === 0) { return 0 }
-
-    const balanceValue = balanceValueArray[0].value.uint
-
-    this.tokenBalances[hotWalletAddress] = balanceValue
-    return balanceValue
-  }
-
-  async getOptedInAppsFromBlockchain(hotWalletAddress) {
-    try {
-      const hwAccountDetails = await this.algoChain.algoClientIndexer.lookupAccountByID(hotWalletAddress).do()
-      return hwAccountDetails.account['apps-local-state'].filter(app => app.deleted == false).map(app => app.id)
-    } catch (err) {
-      return []
-    }
-  }
-
-  async getOptedInAppsForHotWallet(hotWalletAddress) {
-    if (hotWalletAddress in this.optedInApps) { return this.optedInApps[hotWalletAddress] }
-
-    await this.connect()
-    const optedInApps = await this.getOptedInAppsFromBlockchain(hotWalletAddress)
-    if (optedInApps.length > 0) {
-      this.optedInApps[hotWalletAddress] = optedInApps
-    } else {
-      console.log(`The Hot Wallet either has zero algos balance or doesn't opted-in to any app. Please send ALGOs to the wallet ${hotWalletAddress}`)
-    }
-
-    return optedInApps
-  }
-
-  async getAlgoBalanceForHotWallet(hotWalletAddress) {
-    if (hotWalletAddress in this.algoBalances) { return this.algoBalances[hotWalletAddress] }
-
-    await this.connect()
-    const blockchainBalance = await this.algoChain.fetchBalance(hotWalletAddress, chainjs.HelpersAlgorand.toAlgorandSymbol('algo'))
-    this.algoBalances[hotWalletAddress] = new BigNumber(blockchainBalance.balance)
-
-    return this.algoBalances[hotWalletAddress]
-  }
-
-  async isOptedInToCurrentApp(hotWalletAddress) {
-    const optedInApps = await this.getOptedInAppsForHotWallet(hotWalletAddress)
-    return optedInApps.includes(this.envs.optInApp)
-  }
-
-  async enoughAlgoBalanceToSendTransaction(hotWalletAddress) {
-    const algoBalance = await this.getAlgoBalanceForHotWallet(hotWalletAddress)
-    return algoBalance.isLessThan(new BigNumber(0.001)) // 1000 microalgos
-  }
-
-  async positiveTokenBalance(hotWalletAddress) {
-    const tokenBalance = await this.getTokenBalance(hotWalletAddress)
-    return tokenBalance > 0
-  }
-
-  async optInToApp(hotWallet, appToOptIn) {
-    await this.connect()
-
-    const transaction = await this.algoChain.new.Transaction()
-    const txn = {
-      from: hotWallet.address,
-      note: 'Opt-in from a HotWallet',
-      appIndex: appToOptIn,
-    }
-    const action = await this.algoChain.composeAction(chainjs.ModelsAlgorand.AlgorandChainActionType.AppOptIn, txn)
-    transaction.actions = [action]
-    await transaction.prepareToBeSigned()
-    await transaction.validate()
-    await transaction.sign([chainjs.HelpersAlgorand.toAlgorandPrivateKey(hotWallet.secretKey())])
-
-    try {
-      const tx_result = await transaction.send(chainjs.Models.ConfirmType.After001)
-      console.log(`Opt-in was successfully sent to blockchain, tx hash: ${tx_result.transactionId}`)
-      return tx_result
-    } catch (err) {
-      console.error(err)
-      return {}
-    }
-  }
-
-  async isTransactionValid(transaction, hotWalletAddress) {
-    if (typeof transaction !== 'string') { return { valid: false } }
-
-    try {
-      transaction = JSON.parse(transaction)
-
-      // Is AppIndex for the current App?
-      if (transaction.appIndex !== this.envs.optInApp) {
-        const errorMessage = "The transaction is not for configured App."
-        console.log(errorMessage)
-        return { valid: false, markAs: "failed", error: errorMessage }
-      }
-
-      // Checks for transfer transaction
-      if (transaction.appArgs[0] === "0x7472616e73666572") {
-        const txTokensAmount = parseInt(transaction.appArgs[1])
-
-        if (this.envs.maxAmountForTransfer > 0 && txTokensAmount > this.envs.maxAmountForTransfer) {
-          const errorMessage = `The transaction has too big amount for transfer (${txTokensAmount}). Max amount is ${this.envs.maxAmountForTransfer}`
-          console.log(errorMessage)
-          return { valid: false, markAs: "failed", error: errorMessage }
-        }
-
-        const hwTokensBalance = await this.getTokenBalance(hotWalletAddress)
-        if (hwTokensBalance < txTokensAmount) {
-          const errorMessage = `The Hot Wallet has insufficient tokens to transfer (${hwTokensBalance} < ${txTokensAmount})`
-          console.log(errorMessage)
-          return { valid: false, markAs: "cancelled", error: errorMessage }
-        }
-      }
-      return { valid: true }
-    } catch (err) {
-      return { valid: false, markAs: "failed", error: `Unknown error: ${err}` }
-    }
-  }
-}
+const AlgorandBlockchain = require("./blockchains/AlgorandBlockchain").AlgorandBlockchain
 exports.AlgorandBlockchain = AlgorandBlockchain
-
-class ComakeryApi {
-  constructor(envs) {
-    this.envs = envs
-  }
-
-  async registerHotWallet(wallet) {
-    const registerHotWalletUrl = `${this.envs.comakeryServerUrl}/api/v1/projects/${this.envs.projectId}/hot_wallet_addresses`
-    const params = { body: { data: { hot_wallet: { address: wallet.address } } } }
-    const config = { headers: { "API-Transaction-Key": this.envs.projectApiKey } }
-
-    try {
-      return await axios.post(registerHotWalletUrl, params, config)
-    } catch (error) {
-      this.logError('registerHotWallet', error)
-      return {}
-    }
-  }
-
-  async getNextTransactionToSign(hotWalletAddress) {
-    const blockchainTransactionsUrl = `${this.envs.comakeryServerUrl}/api/v1/projects/${this.envs.projectId}/blockchain_transactions`
-    const params = { body: { data: { transaction: { source: hotWalletAddress } } } }
-    const config = { headers: { "API-Transaction-Key": this.envs.projectApiKey } }
-
-    try {
-      const res = await axios.post(blockchainTransactionsUrl, params, config)
-      if (res.status == 201) {
-        return res.data
-      } else {
-        return {}
-      }
-    } catch (error) {
-      this.logError('getNextTransactionToSign', error)
-      return {}
-    }
-  }
-
-  async updateTransactionHash(blockchainTransaction) {
-    const blockchainTransactionsUrl = `${this.envs.comakeryServerUrl}/api/v1/projects/${this.envs.projectId}/blockchain_transactions/${blockchainTransaction.id}`
-    const params = { body: { data: { transaction: { tx_hash: blockchainTransaction.txHash } } } }
-    const config = { headers: { "API-Transaction-Key": this.envs.projectApiKey } }
-
-    try {
-      const res = await axios.put(blockchainTransactionsUrl, params, config)
-      if (res.status == 200) {
-        console.log("Transaction hash was successfully updated")
-        return res.data
-      } else {
-        return {}
-      }
-    } catch (error) {
-      this.logError('updateTransactionHash', error)
-      return {}
-    }
-  }
-
-  async cancelTransaction(blockchainTransaction, errorMessage, status ) {
-    const blockchainTransactionsUrl = `${this.envs.comakeryServerUrl}/api/v1/projects/${this.envs.projectId}/blockchain_transactions/${blockchainTransaction.id}`
-    let transactionParams = { status_message: errorMessage }
-    if (status == "failed") { transactionParams["failed"] = true }
-    const params = { body: { data: { transaction: transactionParams } } }
-    const config = { data: params, headers: { "API-Transaction-Key": this.envs.projectApiKey } }
-
-    try {
-      const res = await axios.delete(blockchainTransactionsUrl, config)
-      if (res.status == 200) {
-        console.log(`The transaction has been marked as ${status}`)
-        return res.data
-      } else {
-        return {}
-      }
-    } catch (error) {
-      this.logError('cancelTransaction', error)
-      return {}
-    }
-  }
-
-  logError(functionName, error) {
-    console.error(`${functionName} API call failed with:\n`)
-
-    if (error.response) {
-      console.error(
-        `${error.response.status} (${error.response.statusText}) data:\n`,
-        error.response.data
-      )
-    } else {
-      console.error(`${functionName} produced an unknown error on API call`)
-    }
-  }
-}
+const ComakeryApi = require("./ComakeryApi").ComakeryApi
 exports.ComakeryApi = ComakeryApi
 
 exports.checkAllVariablesAreSet = function checkAllVariablesAreSet(envs) {
